+++ src/gsee/__init__.py
@@ -1,4 +1,4 @@
-from gsee import brl_model, pv
+from gsee import api, pv
 from gsee._version import __version__
 
-__all__ = ["__version__", "brl_model", "pv"]
+__all__ = ["__version__", "api", "pv"]